+++ routes/dev/instagram.js
@@ -1,7 +1,9 @@
 const { Router } = require('express');
 const router = Router();
-const Instagram = require('../../models/Instagram');
+const Instagram = require('../../models/instagram/instagramProfiles');
+const InstaPost = require('../../models/instagram/instagramPosts');
 const auth = require('../../middleware/auth');
+const { uploadFile } = require('../../helpers/storage');
 
 router.get('/list', async (req, res) => {
     try {
@@ -14,7 +16,7 @@ router.get('/list', async (req, res) => {
 
 router.get('/account/:name', async (req, res) => {
     try {
-        const account = await Instagram.findOne({ name: req.params.name }).select('name stats countSubscribe countUnSubscribe tagLikes countLikes active tested private user');
+        const account = await Instagram.findOne({ name: req.params.name }).select('name stats countSubscribe countUnSubscribe tagLikes countLikes active tested private user title');
         const userId = req.user ? req.user.id : '';
         const canEdit = JSON.parse(JSON.stringify(account)).user === userId;
 
@@ -22,11 +24,14 @@ router.get('/account/:name', async (req, res) => {
             return res.status(422).json({ message: 'not account' });
         }
 
+        const posts = await InstaPost.find({ profileId: account._id });
+
         res.status(201).json({
             message: 'ok',
             account: {
                 canEdit,
-                ...JSON.parse(JSON.stringify(account))
+                ...JSON.parse(JSON.stringify(account)),
+                posts
             }
         });
     } catch (e) {
@@ -40,7 +45,7 @@ router.post('/add', auth, async (req, res) => {
         name,
         password,
         countLikes,
-        tagLikes: tagLikes.split(','),
+        tagLikes: tagLikes.replace(/\s+/g, '').split(','),
         user: req.user.id
     });
 
@@ -62,7 +67,7 @@ router.put('/edit/:id', auth, async (req, res) => {
 
     account.password = password ? password : account.password;
     account.countLikes = countLikes ? countLikes : account.countLikes;
-    account.tagLikes = tagLikes.split(',').length > 0 ? tagLikes.split(',') : account.tagLikes;
+    account.tagLikes = tagLikes.split(',').length > 0 ? tagLikes.replace(/\s+/g, '').split(',') : account.tagLikes;
     account.active = active;
     account.tested = tested;
     account.private = privates;
@@ -90,6 +95,46 @@ router.delete('/delete/:id', async (req, res) => {
         });
     } catch (e) {
         return res.status(422).json({ message: 'not account' });
+    }
+});
+
+router.post('/post/', auth, async (req, res) => {
+    const { title, description, imageUrl, tags, datePublish, profileId } = req.body;
+    const profile = await Instagram.findById(profileId);
+
+    if (profile.user != req.user.id) {
+        return res.status(422).json({ message: 'not authorized' });
+    }
+
+    const image = 'https://storage.indiro.ru/instagram/' + imageUrl.split('/').pop();
+    await uploadFile(imageUrl, 'instagram');
+
+    const newPost = new InstaPost({
+        title,
+        description,
+        imageUrl: image,
+        tags,
+        datePublish,
+        profileId
+    });
+
+    try {
+        await newPost.save();
+        res.status(201).json({ message: 'post added' });
+    } catch (e) {
+        return res.status(422).json({ message: 'post not added' });
+    }
+});
+
+router.delete('/post/:id', async (req, res) => {
+    try {
+        await InstaPost.findOneAndDelete({ _id: req.params.id });
+
+        res.status(201).json({
+            message: 'post was deleted'
+        });
+    } catch (e) {
+        return res.status(422).json({ message: 'not post' });
     }
 });