+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 #CHANGE LOG
 
+## [1.0.5] - 2021-02-13
+### Added
+- Add instagram posts functional
+- Add instagram posts shared cron
+
 ## [1.0.4] - 2021-02-08
 ### Added
 - Add tags functional